GHASH test cleanup
