+++ circuits/aes-gcm/utils.circom
@@ -5,6 +5,100 @@ include "circomlib/circuits/mux1.circom";
 include "circomlib/circuits/bitify.circom";
 include "circomlib/circuits/gates.circom";
 
+// n is the number of bytes to convert to bits
+template BytesToBits(n_bytes) {
+    signal input in[n_bytes];
+    signal output out[n_bytes*8];
+    component num2bits[n_bytes];
+    for (var i = 0; i < n_bytes; i++) {
+        num2bits[i] = Num2Bits(8);
+        num2bits[i].in <== in[i];
+        for (var j = 7; j >=0; j--) {
+            out[i*8 + j] <== num2bits[i].out[7 -j];
+        }
+    }
+}
+
+// n is the number of bytes we want
+template BitsToBytes(n) {
+    signal input in[n*8];
+    signal output out[n];
+    component bits2num[n];
+    for (var i = 0; i < n; i++) {
+        bits2num[i] = Bits2Num(8);
+        for (var j = 0; j < 8; j++) {
+            bits2num[i].in[7 - j] <== in[i*8 + j];
+        }
+        out[i] <== bits2num[i].out;
+    }
+}
+
+// XORs two arrays of bits
+template XorBits(){
+        signal input a[8];
+        signal input b[8];
+        signal output out[8];
+
+    component xor[8];
+    for (var i = 0; i < 8; i++) {
+        xor[i] = XOR();
+        xor[i].a <== a[i];
+        xor[i].b <== b[i];
+        out[i] <== xor[i].out;
+    }
+}
+
+// XORs two bytes
+template XorByte(){
+        signal input a;
+        signal input b;
+        signal output out;
+
+        component abits = Num2Bits(8);
+        abits.in <== a;
+
+        component bbits = Num2Bits(8);
+        bbits.in <== b;
+
+        component XorBits = XorBits();
+        XorBits.a <== abits.out;
+        XorBits.b <== bbits.out;
+
+        component num = Bits2Num(8);
+        num.in <== XorBits.out;
+
+        out <== num.out;
+}
+
+// XOR n bytes
+template XORBLOCK(n_bytes){
+    signal input a[n_bytes];
+    signal input b[n_bytes];
+    signal output out[n_bytes];
+
+    component xorByte[n_bytes];
+    for (var i = 0; i < n_bytes; i++) {
+        xorByte[i] = XorByte();
+        xorByte[i].a <== a[i];
+        xorByte[i].b <== b[i];
+        out[i] <== xorByte[i].out;
+    }
+}
+
+// multiplexer for arrays of length n
+template ArrayMux(n) {
+    signal input a[n];      // First input array
+    signal input b[n];      // Second input array
+    signal input sel;       // Selector signal (0 or 1)
+    signal output out[n];   // Output array
+
+    for (var i = 0; i < n; i++) {
+        // If sel = 0, out[i] = a[i]
+        // If sel = 1, out[i] = b[i]
+        out[i] <== (b[i] - a[i]) * sel + a[i];
+    }
+}
+
 // parse LE bits to int
 template ParseLEBytes64() {
     signal input in[64];
