+++ circuits/test/common/utils.test.ts
@@ -1,5 +1,5 @@
 import { WitnessTester } from "circomkit";
-import { bitArrayToHex, circomkit, hexToBitArray } from ".";
+import { bitArrayToHex, circomkit, hexByteToBigInt, hexToBitArray } from ".";
 import { assert } from "chai";
 
 describe("reverse_byte_array", () => {
@@ -111,5 +111,110 @@ describe("ParseBytesLE", () => {
     const result = _result.out as number;
 
     assert.equal(result, expected, "parse incorrect");
+  });
+});
+
+describe("ArrayMux", () => {
+  let circuit: WitnessTester<["a", "b", "sel"], ["out"]>;
+
+  before(async () => {
+    circuit = await circomkit.WitnessTester("XORBLOCK", {
+      file: "aes-gcm/utils",
+      template: "ArrayMux",
+      params: [16]
+    });
+    console.log("#constraints:", await circuit.getConstraintCount());
+  });
+  // msb is 1 so we xor the first byte with 0xE1
+  it("Should Compute selector mux Correctly", async () => {
+    let a = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
+    let b = [0xE1, 0xE1, 0xE1, 0xE1, 0xE1, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01];
+    let sel = 0x00;
+    let expected = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
+    await circuit.expectPass({ a: a, b: b, sel: sel }, { out: expected });
+  });
+
+  it("Should Compute block XOR Correctly", async () => {
+    let a = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
+    let b = [0xE1, 0xE1, 0xE1, 0xE1, 0xE1, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01];
+    let sel = 0x01;
+    let expected = [0xE1, 0xE1, 0xE1, 0xE1, 0xE1, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01];
+    await circuit.expectPass({ a: a, b: b, sel: sel }, { out: expected });
+  });
+
+});
+describe("XORBLOCK", () => {
+  let circuit: WitnessTester<["a", "b"], ["out"]>;
+
+  before(async () => {
+    circuit = await circomkit.WitnessTester("XORBLOCK", {
+      file: "aes-gcm/utils",
+      template: "XORBLOCK",
+      params: [16]
+    });
+    console.log("#constraints:", await circuit.getConstraintCount());
+  });
+  // msb is 1 so we xor the first byte with 0xE1
+  it("Should Compute block XOR Correctly", async () => {
+    let inputa = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
+    let inputb = [0xE1, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01];
+    const expected = [0xE1, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01];
+    await circuit.expectPass({ a: inputa, b: inputb }, { out: expected });
+  });
+});
+
+describe("ToBytes", () => {
+  let circuit: WitnessTester<["in"], ["out"]>;
+
+  before(async () => {
+    circuit = await circomkit.WitnessTester("bytesToBits", {
+      file: "aes-gcm/utils",
+      template: "BitsToBytes",
+      params: [1]
+    });
+    console.log("#constraints:", await circuit.getConstraintCount());
+  });
+
+  it("Should Compute bytesToBits Correctly", async () => {
+    let input = hexToBitArray("0x01");
+    const expected = hexByteToBigInt("0x01");
+    // console.log("expected", expected);
+    const _res = await circuit.compute({ in: input }, ["out"]);
+    // console.log("res:", _res.out);
+    assert.deepEqual(_res.out, expected);
+  });
+  it("Should Compute bytesToBits Correctly", async () => {
+    let input = hexToBitArray("0xFF");
+    const expected = hexByteToBigInt("0xFF");
+    // console.log("expected", expected);
+    const _res = await circuit.compute({ in: input }, ["out"]);
+    // console.log("res:", _res.out);
+    assert.deepEqual(_res.out, expected);
+  });
+});
+
+describe("ToBits", () => {
+  let circuit: WitnessTester<["in"], ["out"]>;
+
+  before(async () => {
+    circuit = await circomkit.WitnessTester("bytesToBits", {
+      file: "aes-gcm/utils",
+      template: "BytesToBits",
+      params: [2]
+    });
+    console.log("#constraints:", await circuit.getConstraintCount());
+  });
+
+  it("Should Compute bytesToBits Correctly", async () => {
+    let input = [0x01, 0x00];
+    const expected = hexToBitArray("0x0100");
+    // console.log("expected", expected);
+    const _res = await circuit.expectPass({ in: input }, { out: expected });
+  });
+  it("Should Compute bytesToBits Correctly", async () => {
+    let input = [0xFF, 0x00];
+    const expected = hexToBitArray("0xFF00");
+    // console.log("expected", expected);
+    const _res = await circuit.expectPass({ in: input }, { out: expected });
   });
 });
